Clean up a big mess of .nfs file put into CVS!

+++ Library/src/HTMLGen.c
@@ -96,7 +96,7 @@ PRIVATE void HTMLGen_put_character ARGS2(HTStructured *, me, char, c)
 	    char delims[5];
 	    char * p;
 	    strcpy(delims, ",;:.");		/* @@ english bias */
-	    p = strchr(delims, me->write_pointer[-2]);
+	    p = strchr(delims, (int)me->write_pointer[-2]);
 	    if (p) new_cleanness = p - delims + 2;
 	}
 	if (new_cleanness >= me->cleanness) {
